pyproject.toml
[project]
name = "autonomousclinicalsystem"
version = "0.1.0"
description = "AI-Powered Clinical Decision Support System"
authors = [
    {name = "razbi", email = "razhaim101@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langchain>=0.2.0,<0.3.0",
    "langchain-community>=0.2.0,<0.3.0",
    "langchain-openai>=0.1.0,<0.2.0",
    "langchain-pinecone>=0.1.0,<0.2.0",
    "langgraph>=0.1.0,<0.2.0",
    "pinecone-client>=4.0.0,<6.0.0",
    "tiktoken>=0.7.0,<1.0.0",
    "fastapi>=0.111.0,<0.120.0",
    "uvicorn>=0.30.0,<0.40.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.7.0,<3.0.0",
    "numpy<2.0.0",
    "matplotlib>=3.9.0,<4.0.0",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "ragas (>=0.4.3,<0.5.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)",
    "ultralytics (>=8.4.14,<9.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false